Move nested classes to own files
